feat(topology/core): add node/pool topology code to core-agent
