+++ control-plane/agents/src/bin/core/controller/scheduling/mod.rs
@@ -6,9 +6,9 @@ pub(crate) mod volume;
 mod volume_policy;
 
 use crate::controller::scheduling::{
-    nexus::{GetPersistedNexusChildrenCtx, GetSuitableNodesContext},
-    resources::{ChildItem, NodeItem, PoolItem, ReplicaItem},
-    volume::{GetSuitablePoolsContext, ReplicaResizePoolsContext, VolumeReplicasForNexusCtx},
+    nexus::GetPersistedNexusChildrenCtx,
+    resources::{ChildItem, PoolItem, ReplicaItem},
+    volume::{ReplicaResizePoolsContext, VolumeReplicasForNexusCtx},
 };
 use std::{cmp::Ordering, collections::HashMap, future::Future};
 use weighted_scoring::{Criteria, Value, ValueGrading, WeightedScore};
@@ -151,71 +151,6 @@ impl SortBuilder {
         }
         let (score_a, score_b) = weighted_score.score().unwrap();
         score_b.cmp(&score_a)
-    }
-}
-
-/// Filter nodes used for replica creation
-pub(crate) struct NodeFilters {}
-impl NodeFilters {
-    /// Should only attempt to use online nodes for pools.
-    pub(crate) fn online_for_pool(_request: &GetSuitablePoolsContext, item: &PoolItem) -> bool {
-        item.node.is_online()
-    }
-    /// Should only attempt to use allowed nodes (by the topology).
-    pub(crate) fn allowed(request: &GetSuitablePoolsContext, item: &PoolItem) -> bool {
-        request.allowed_nodes().is_empty() || request.allowed_nodes().contains(&item.pool.node)
-    }
-    /// Should only attempt to use nodes not currently used by the volume.
-    /// When moving a replica the current replica node is allowed to be reused for a different pool.
-    pub(crate) fn unused(request: &GetSuitablePoolsContext, item: &PoolItem) -> bool {
-        if let Some(moving) = request.move_repl() {
-            if moving.node() == &item.pool.node && moving.pool() != &item.pool.id {
-                return true;
-            }
-        }
-        let registry = request.registry();
-        let used_nodes = registry.specs().volume_data_nodes(&request.uuid);
-        !used_nodes.contains(&item.pool.node)
-    }
-    /// Should only attempt to use nodes which are not cordoned.
-    pub(crate) fn cordoned_for_pool(request: &GetSuitablePoolsContext, item: &PoolItem) -> bool {
-        let registry = request.registry();
-        !registry
-            .specs()
-            .cordoned_nodes()
-            .into_iter()
-            .any(|node_spec| node_spec.id() == &item.pool.node)
-    }
-
-    /// Should only attempt to use online nodes.
-    pub(crate) fn online(_request: &GetSuitableNodesContext, item: &NodeItem) -> bool {
-        item.node_wrapper().is_online()
-    }
-
-    /// Should only attempt to use nodes which are not cordoned.
-    pub(crate) fn cordoned(request: &GetSuitableNodesContext, item: &NodeItem) -> bool {
-        let registry = request.registry();
-        !registry
-            .specs()
-            .cordoned_nodes()
-            .into_iter()
-            .any(|node_spec| node_spec.id() == item.node_wrapper().id())
-    }
-
-    /// Should only attempt to use node where current target is not present.
-    pub(crate) fn current_target(request: &GetSuitableNodesContext, item: &NodeItem) -> bool {
-        if let Some(target) = request.target() {
-            target.node() != item.node_wrapper().id()
-        } else {
-            true
-        }
-    }
-    /// Should only attempt to use node where there are no targets for the current volume.
-    pub(crate) fn no_targets(request: &GetSuitableNodesContext, item: &NodeItem) -> bool {
-        let volume_targets = request.registry().specs().volume_nexuses(&request.uuid);
-        !volume_targets
-            .into_iter()
-            .any(|n| &n.lock().node == item.node_wrapper().id())
     }
 }
 
@@ -413,24 +348,5 @@ impl AddReplicaSorters {
                 }
             }
         }
-    }
-}
-
-/// Sort nodes to pick the best choice for nexus target.
-pub(crate) struct NodeSorters {}
-impl NodeSorters {
-    /// Sort nodes by the number of active nexus present per node.
-    /// The lesser the number of active nexus on a node, the more would be its selection priority.
-    /// In case this is a Affinity Group, then it would be spread on basis of number of ag targets
-    /// and then on basis of total targets on equal.
-    pub(crate) fn number_targets(a: &NodeItem, b: &NodeItem) -> std::cmp::Ordering {
-        a.ag_nexus_count()
-            .cmp(&b.ag_nexus_count())
-            .then_with(|| a.ag_preferred().cmp(&b.ag_preferred()).reverse())
-            .then_with(|| {
-                a.node_wrapper()
-                    .nexus_count()
-                    .cmp(&b.node_wrapper().nexus_count())
-            })
     }
 }
